move PackageCopyDialog to PackageDialog/Copy
