+++ catalog/app/containers/Bucket/PackageDialog/State/index.tsx
@@ -1,3 +1,2 @@
 export * from './State'
 export * from './Provider'
-export { isPackageHandle } from './manifest'
