+++ catalog/app/containers/Bucket/PackageDialog/index.ts
@@ -1,12 +1,9 @@
-export { default as DialogError } from './DialogError'
-export { default as DialogLoading } from './DialogLoading'
-export { default as DialogSuccess } from './DialogSuccess'
 export * as Inputs from './Inputs'
 export * from './Layout'
-export * from './Manifest'
 export * from './PackageCreationForm'
 export * from './Skeleton'
-export { useContext, Provider, isPackageHandle } from './State'
+export { useContext, Provider } from './State'
 export { default as SubmitSpinner } from './SubmitSpinner'
 export * from './Uploads'
 export { default as useInitialActions } from './actions'
+export { default as Copy } from './Copy'
